fix(ec2_ebs_snapshots_exists): rename check and add tests
