+++ prowler/providers/aws/services/ec2/ec2_ebs_volume_snapshots_exists/ec2_ebs_volume_snapshots_exists.py
@@ -2,7 +2,7 @@ from prowler.lib.check.models import Check, Check_Report_AWS
 from prowler.providers.aws.services.ec2.ec2_client import ec2_client
 
 
-class ec2_ebs_snapshots_exists(Check):
+class ec2_ebs_volume_snapshots_exists(Check):
     def execute(self):
         findings = []
         for volume in ec2_client.volumes:
@@ -12,9 +12,13 @@ class ec2_ebs_snapshots_exists(Check):
             report.resource_id = volume.id
             report.resource_arn = volume.arn
             report.resource_tags = volume.tags
-            report.status_extended = f"Snapshots not found to EBS volume {volume.id}"
+            report.status_extended = (
+                f"Snapshots not found for the EBS volume {volume.id}."
+            )
             if ec2_client.volumes_with_snapshots.get(volume.id, False):
                 report.status = "PASS"
-                report.status_extended = f"Snapshots found to EBS volume {volume.id}"
+                report.status_extended = (
+                    f"Snapshots found for the EBS volume {volume.id}."
+                )
             findings.append(report)
         return findings
